papaya

Struct HashSetRef

pub struct HashSetRef<'set, K, S, G> { /* private fields */ }

A pinned reference to a HashSet.

This type is created with HashSet::pin and can be used to easily access a HashSet without explicitly managing a guard. See the crate-level documentation for details.

Implementations

impl<'set, K, S, G> HashSetRef<'set, K, S, G>
where K: Hash + Eq, S: BuildHasher, G: Guard,

pub fn set(&self) -> &'set HashSet<K, S>

Returns a reference to the inner HashSet.

pub fn len(&self) -> usize

Returns the number of entries in the set.

See HashSet::len for details.

pub fn is_empty(&self) -> bool

Returns true if the set is empty. Otherwise returns false.

See HashSet::is_empty for details.

pub fn contains<Q>(&self, key: &Q) -> bool
where Q: Equivalent<K> + Hash + ?Sized,

Returns true if the set contains a value for the specified key.

See HashSet::contains for details.

pub fn get<Q>(&self, key: &Q) -> Option<&K>
where Q: Equivalent<K> + Hash + ?Sized,

Returns a reference to the value corresponding to the key.

See HashSet::get for details.

pub fn insert(&self, key: K) -> bool

Inserts a key-value pair into the set.

See HashSet::insert for details.

pub fn remove<Q>(&self, key: &Q) -> bool
where Q: Equivalent<K> + Hash + ?Sized,

Removes a key from the set, returning the value at the key if the key was previously in the set.

See HashSet::remove for details.

pub fn clear(&self)

Clears the set, removing all values.

See HashSet::clear for details.

pub fn retain<F>(&mut self, f: F)
where F: FnMut(&K) -> bool,

Retains only the elements specified by the predicate.

See HashSet::retain for details.

pub fn reserve(&self, additional: usize)

Tries to reserve capacity for additional more elements to be inserted in the set.

See HashSet::reserve for details.

pub fn iter(&self) -> Iter<'_, K, G>

An iterator visiting all values in arbitrary order. The iterator element type is (&K, &V).

See HashSet::iter for details.